renamed gcrhinterface.h --> RedhawkGCInterface
