+++ src/coreclr/nativeaot/Runtime/CachedInterfaceDispatch.cpp
@@ -26,7 +26,7 @@
 #include "allocheap.h"
 #include "rhbinder.h"
 #include "ObjectLayout.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "shash.h"
 #include "TypeManager.h"
 #include "RuntimeInstance.h"
