+++ src/coreclr/nativeaot/Runtime/EHHelpers.cpp
@@ -7,7 +7,7 @@
 #include "daccess.h"
 #include "rhassert.h"
 #include "slist.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "shash.h"
 #include "TypeManager.h"
 #include "varint.h"
