+++ src/coreclr/nativeaot/Runtime/FinalizerHelpers.cpp
@@ -9,7 +9,7 @@
 #include "gcheaputilities.h"
 
 #include "slist.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "RuntimeInstance.h"
 #include "shash.h"
