+++ src/coreclr/nativeaot/Runtime/GcStressControl.cpp
@@ -14,7 +14,7 @@
 #include "holder.h"
 #include "Crst.h"
 #include "RhConfig.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "slist.h"
 #include "varint.h"
 #include "regdisplay.h"
