+++ src/coreclr/nativeaot/Runtime/MiscHelpers.cpp
@@ -18,7 +18,7 @@
 #include "rhbinder.h"
 #include "RuntimeInstance.h"
 #include "regdisplay.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "varint.h"
 #include "StackFrameIterator.h"
 #include "thread.h"
@@ -26,7 +26,7 @@
 #include "threadstore.h"
 #include "threadstore.inl"
 #include "thread.inl"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "shash.h"
 #include "TypeManager.h"
 #include "MethodTable.h"
