+++ src/coreclr/nativeaot/Runtime/RedhawkGCInterface.h
@@ -114,8 +114,6 @@ public:
     static void EnumGcRef(PTR_RtuObjectRef pRef, GCRefKind kind, void * pfnEnumCallback, void * pvCallbackData);
     static void EnumGcRefConservatively(PTR_RtuObjectRef pRef, void* pfnEnumCallback, void* pvCallbackData);
 
-    static void BulkEnumGcObjRef(PTR_RtuObjectRef pRefs, uint32_t cRefs, void * pfnEnumCallback, void * pvCallbackData);
-
     static void EnumGcRefs(ICodeManager * pCodeManager,
                            MethodInfo * pMethodInfo,
                            PTR_VOID safePointAddress,
