+++ src/coreclr/nativeaot/Runtime/RestrictedCallouts.cpp
@@ -15,7 +15,7 @@
 #include "rhassert.h"
 #include "slist.h"
 #include "holder.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "shash.h"
 #include "rhbinder.h"
 #include "Crst.h"
