+++ src/coreclr/nativeaot/Runtime/RuntimeInstance.cpp
@@ -18,7 +18,7 @@
 #include "thread.h"
 #include "threadstore.h"
 #include "threadstore.inl"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "shash.h"
 #include "TypeManager.h"
 #include "MethodTable.h"
