+++ src/coreclr/nativeaot/Runtime/gcenv.h
@@ -35,7 +35,7 @@
 #include "ObjectLayout.h"
 #include "PalRedhawkCommon.h"
 #include "PalRedhawk.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "gcenv.interlocked.inl"
 
 #include "slist.h"
