+++ src/coreclr/nativeaot/Runtime/gcrhenv.cpp
@@ -17,7 +17,7 @@
 
 #include "RestrictedCallouts.h"
 
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 
 #include "slist.h"
 #include "varint.h"
@@ -66,7 +66,7 @@ RhConfig * g_pRhConfig = &g_sRhConfig;
 //
 // The rest of Redhawk needs to be able to talk to the GC/HandleTable code (to initialize it, allocate
 // objects etc.) without pulling in the entire adaptation layer provided by this file and gcrhenv.h. To this
-// end the rest of Redhawk talks to us via a simple interface described in gcrhinterface.h. We provide the
+// end the rest of Redhawk talks to us via a simple interface described in RedhawkGCInterface.h. We provide the
 // implementation behind those APIs here.
 //
 
@@ -291,9 +291,6 @@ void MethodTable::InitializeAsGcFreeType()
 #endif // !DACCESS_COMPILE
 
 extern void GcEnumObject(PTR_OBJECTREF pObj, uint32_t flags, EnumGcRefCallbackFunc * fnGcEnumRef, EnumGcRefScanContext * pSc);
-extern void GcEnumObjectConservatively(PTR_PTR_Object ppObj, EnumGcRefCallbackFunc* fnGcEnumRef, EnumGcRefScanContext* pSc);
-extern void GcEnumObjectsConservatively(PTR_OBJECTREF pLowerBound, PTR_OBJECTREF pUpperBound, EnumGcRefCallbackFunc * fnGcEnumRef, EnumGcRefScanContext * pSc);
-extern void GcBulkEnumObjects(PTR_OBJECTREF pObjs, DWORD cObjs, EnumGcRefCallbackFunc * fnGcEnumRef, EnumGcRefScanContext * pSc);
 
 struct EnumGcRefContext : GCEnumContext
 {
@@ -330,6 +327,79 @@ void RedhawkGCInterface::EnumGcRefs(ICodeManager * pCodeManager,
                              isActiveStackFrame);
 }
 
+void PromoteCarefully(PTR_PTR_Object obj, uint32_t flags, EnumGcRefCallbackFunc* fnGcEnumRef, EnumGcRefScanContext* pSc)
+{
+    //
+    // Sanity check that the flags contain only these values
+    //
+    assert((flags & ~(GC_CALL_INTERIOR | GC_CALL_PINNED)) == 0);
+
+    //
+    // Sanity check that GC_CALL_INTERIOR FLAG is set
+    //
+    assert(flags & GC_CALL_INTERIOR);
+
+    // If the object reference points into the stack, we
+    // must not promote it, the GC cannot handle these.
+    if (pSc->thread_under_crawl->IsWithinStackBounds(*obj))
+        return;
+
+    fnGcEnumRef(obj, pSc, flags);
+}
+
+void GcEnumObject(PTR_PTR_Object ppObj, uint32_t flags, EnumGcRefCallbackFunc* fnGcEnumRef, EnumGcRefScanContext* pSc)
+{
+    //
+    // Sanity check that the flags contain only these values
+    //
+    assert((flags & ~(GC_CALL_INTERIOR | GC_CALL_PINNED)) == 0);
+
+    // for interior pointers, we optimize the case in which
+    //  it points into the current threads stack area
+    //
+    if (flags & GC_CALL_INTERIOR)
+        PromoteCarefully(ppObj, flags, fnGcEnumRef, pSc);
+    else
+        fnGcEnumRef(ppObj, pSc, flags);
+}
+
+// Scan a contiguous range of memory and report everything that looks like it could be a GC reference as a
+// pinned interior reference. Pinned in case we are wrong (so the GC won't try to move the object and thus
+// corrupt the original memory value by relocating it). Interior since we (a) can't easily tell whether a
+// real reference is interior or not and interior is the more conservative choice that will work for both and
+// (b) because it might not be a real GC reference at all and in that case falsely listing the reference as
+// non-interior will cause the GC to make assumptions and crash quite quickly.
+void GcEnumObjectsConservatively(PTR_PTR_Object ppLowerBound, PTR_PTR_Object ppUpperBound, EnumGcRefCallbackFunc* fnGcEnumRef, EnumGcRefScanContext* pSc)
+{
+    // Only report potential references in the promotion phase. Since we report everything as pinned there
+    // should be no work to do in the relocation phase.
+    if (pSc->promotion)
+    {
+        for (PTR_PTR_Object ppObj = ppLowerBound; ppObj < ppUpperBound; ppObj++)
+        {
+            // Only report values that lie in the GC heap range. This doesn't conclusively guarantee that the
+            // value is a GC heap reference but it's a cheap check that weeds out a lot of spurious values.
+            PTR_Object pObj = *ppObj;
+            if (((PTR_UInt8)pObj >= g_lowest_address) && ((PTR_UInt8)pObj <= g_highest_address))
+                PromoteCarefully(ppObj, GC_CALL_INTERIOR | GC_CALL_PINNED, fnGcEnumRef, pSc);
+        }
+    }
+}
+
+void GcEnumObjectConservatively(PTR_PTR_Object ppObj, EnumGcRefCallbackFunc* fnGcEnumRef, EnumGcRefScanContext* pSc)
+{
+    // Only report potential references in the promotion phase. Since we report everything as pinned there
+    // should be no work to do in the relocation phase.
+    if (pSc->promotion)
+    {
+        // Only report values that lie in the GC heap range. This doesn't conclusively guarantee that the
+        // value is a GC heap reference but it's a cheap check that weeds out a lot of spurious values.
+        PTR_Object pObj = *ppObj;
+        if (((PTR_UInt8)pObj >= g_lowest_address) && ((PTR_UInt8)pObj <= g_highest_address))
+            PromoteCarefully(ppObj, GC_CALL_INTERIOR | GC_CALL_PINNED, fnGcEnumRef, pSc);
+    }
+}
+
 // static
 void RedhawkGCInterface::EnumGcRefsInRegionConservatively(PTR_RtuObjectRef pLowerBound,
                                                           PTR_RtuObjectRef pUpperBound,
@@ -361,12 +431,6 @@ void RedhawkGCInterface::EnumGcRefConservatively(PTR_RtuObjectRef pRef, void* pf
 }
 
 #ifndef DACCESS_COMPILE
-
-// static
-void RedhawkGCInterface::BulkEnumGcObjRef(PTR_RtuObjectRef pRefs, uint32_t cRefs, void * pfnEnumCallback, void * pvCallbackData)
-{
-    GcBulkEnumObjects((PTR_OBJECTREF)pRefs, cRefs, (EnumGcRefCallbackFunc *)pfnEnumCallback, (EnumGcRefScanContext *)pvCallbackData);
-}
 
 // static
 GcSegmentHandle RedhawkGCInterface::RegisterFrozenSegment(void * pSection, size_t allocSize, size_t commitSize, size_t reservedSize)
