+++ src/coreclr/nativeaot/Runtime/gcrhscan.cpp
@@ -10,7 +10,7 @@
 
 #include "PalRedhawkCommon.h"
 
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 
 #include "slist.h"
 #include "varint.h"
@@ -27,7 +27,6 @@
 
 #ifndef DACCESS_COMPILE
 
-void GcEnumObjectsConservatively(PTR_PTR_Object ppLowerBound, PTR_PTR_Object ppUpperBound, EnumGcRefCallbackFunc * fnGcEnumRef, EnumGcRefScanContext * pSc);
 void GcEnumObject(PTR_PTR_Object ppObj, uint32_t flags, EnumGcRefCallbackFunc* fnGcEnumRef, EnumGcRefScanContext* pSc);
 
 /*
@@ -93,84 +92,3 @@ void GCToEEInterface::GcEnumAllocContexts (enum_alloc_context_func* fn, void* pa
 }
 
 #endif //!DACCESS_COMPILE
-
-void PromoteCarefully(PTR_PTR_Object obj, uint32_t flags, EnumGcRefCallbackFunc * fnGcEnumRef, EnumGcRefScanContext * pSc)
-{
-    //
-    // Sanity check that the flags contain only these values
-    //
-    assert((flags & ~(GC_CALL_INTERIOR|GC_CALL_PINNED)) == 0);
-
-    //
-    // Sanity check that GC_CALL_INTERIOR FLAG is set
-    //
-    assert(flags & GC_CALL_INTERIOR);
-
-    // If the object reference points into the stack, we
-    // must not promote it, the GC cannot handle these.
-    if (pSc->thread_under_crawl->IsWithinStackBounds(*obj))
-        return;
-
-    fnGcEnumRef(obj, pSc, flags);
-}
-
-void GcEnumObject(PTR_PTR_Object ppObj, uint32_t flags, EnumGcRefCallbackFunc * fnGcEnumRef, EnumGcRefScanContext * pSc)
-{
-    //
-    // Sanity check that the flags contain only these values
-    //
-    assert((flags & ~(GC_CALL_INTERIOR|GC_CALL_PINNED)) == 0);
-
-    // for interior pointers, we optimize the case in which
-    //  it points into the current threads stack area
-    //
-    if (flags & GC_CALL_INTERIOR)
-        PromoteCarefully (ppObj, flags, fnGcEnumRef, pSc);
-    else
-        fnGcEnumRef(ppObj, pSc, flags);
-}
-
-void GcBulkEnumObjects(PTR_PTR_Object pObjs, uint32_t cObjs, EnumGcRefCallbackFunc * fnGcEnumRef, EnumGcRefScanContext * pSc)
-{
-    PTR_PTR_Object ppObj = pObjs;
-
-    for (uint32_t i = 0; i < cObjs; i++)
-        fnGcEnumRef(ppObj++, pSc, 0);
-}
-
-// Scan a contiguous range of memory and report everything that looks like it could be a GC reference as a
-// pinned interior reference. Pinned in case we are wrong (so the GC won't try to move the object and thus
-// corrupt the original memory value by relocating it). Interior since we (a) can't easily tell whether a
-// real reference is interior or not and interior is the more conservative choice that will work for both and
-// (b) because it might not be a real GC reference at all and in that case falsely listing the reference as
-// non-interior will cause the GC to make assumptions and crash quite quickly.
-void GcEnumObjectsConservatively(PTR_PTR_Object ppLowerBound, PTR_PTR_Object ppUpperBound, EnumGcRefCallbackFunc * fnGcEnumRef, EnumGcRefScanContext * pSc)
-{
-    // Only report potential references in the promotion phase. Since we report everything as pinned there
-    // should be no work to do in the relocation phase.
-    if (pSc->promotion)
-    {
-        for (PTR_PTR_Object ppObj = ppLowerBound; ppObj < ppUpperBound; ppObj++)
-        {
-            // Only report values that lie in the GC heap range. This doesn't conclusively guarantee that the
-            // value is a GC heap reference but it's a cheap check that weeds out a lot of spurious values.
-            PTR_Object pObj = *ppObj;
-            if (((PTR_UInt8)pObj >= g_lowest_address) && ((PTR_UInt8)pObj <= g_highest_address))
-                PromoteCarefully(ppObj, GC_CALL_INTERIOR | GC_CALL_PINNED, fnGcEnumRef, pSc);
-        }
-    }
-}
-
-void GcEnumObjectConservatively(PTR_PTR_Object ppObj, EnumGcRefCallbackFunc* fnGcEnumRef, EnumGcRefScanContext* pSc)
-{
-    // Only report potential references in the promotion phase. Since we report everything as pinned there
-    // should be no work to do in the relocation phase.
-    if (pSc->promotion)
-    {
-        // Only report values that lie in the GC heap range. This doesn't conclusively guarantee that the
-        // value is a GC heap reference but it's a cheap check that weeds out a lot of spurious values.
-        PTR_Object pObj = *ppObj;
-        if (((PTR_UInt8)pObj >= g_lowest_address) && ((PTR_UInt8)pObj <= g_highest_address))
-            PromoteCarefully(ppObj, GC_CALL_INTERIOR | GC_CALL_PINNED, fnGcEnumRef, pSc);
-    }
-}
