+++ src/coreclr/nativeaot/Runtime/interoplibinterface_objc.cpp
@@ -10,7 +10,7 @@
 #include "rhassert.h"
 #include "slist.h"
 #include "holder.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "shash.h"
 #include "rhbinder.h"
 #include "Crst.h"
