+++ src/coreclr/nativeaot/Runtime/portable.cpp
@@ -12,7 +12,7 @@
 #include "rhassert.h"
 
 #include "slist.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "shash.h"
 #include "varint.h"
 #include "holder.h"
