+++ src/coreclr/nativeaot/Runtime/startup.cpp
@@ -8,7 +8,7 @@
 #include "PalRedhawk.h"
 #include "rhassert.h"
 #include "slist.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "varint.h"
 #include "regdisplay.h"
 #include "StackFrameIterator.h"
