+++ src/coreclr/nativeaot/Runtime/stressLog.cpp
@@ -21,7 +21,7 @@
 #include "Crst.h"
 #include "rhassert.h"
 #include "slist.h"
-#include "gcrhinterface.h"
+#include "RedhawkGCInterface.h"
 #include "varint.h"
 #include "regdisplay.h"
 #include "StackFrameIterator.h"
